+++ datasets/__init__.py
@@ -3,6 +3,8 @@ from .datamodule import BaseDataModule, StandardSplitDataModule, KFoldDataModule
 from .transforms import build_transforms
 from .polyp.clinicdb import ClinicDB
 from .polyp.colondb import ColonDB
+from .busi import BUSI
+from .isic18 import ISIC18
 
 __all__ = [
     "MedicalSegmentationDataset",
@@ -13,4 +15,6 @@ __all__ = [
     "build_transforms",
     "ClinicDB",
     "ColonDB",
+    "BUSI",
+    "ISIC18",
 ]